Fix create_vm integartion suite
This code fixes a problem in create_vm integration test, it always fails
with:

`Creating VirtualGuest from SoftLayer client: * Hostname for the
computing instance is required and cannot be empty`.

JSON template specified hostname as value of 'hostname' field but the
rest of the code expected it in 'VmPrefix' field. Finaly it should be
asigned to 'Hostname' field, so this fix harmonize naming by introducing
'Hostname' field in Go code.
Also value of 'hostname' in JSON was invalid, RFC doesn't allow to have
'_' in host names, SoftLayer API rejects such requests.
Small refactoring in spec code.